Update aacgm C library to version 2.4
Not an exact 1-to-1 copy because of mixed space/tab indenting in the
original source.

Procedure: First, diff'ed ignoring white space changes, then applied and
fixed some left-over space indents with tabs to match the original
(version 2.2) indenting.

In some places the code is still indented with two spaces instead of
tabs. We changed only the lines that were modified and kept the others
as they were.