Return the converted altitude as well
As of version 2.4 of the C aacgm library, the converted altitude
is supposed to be useful. We promote the returned altitude to the
python wrapper function so that people can use it if they want to.

Updates the function's inline documentation, the command line interface,
and the tests (standalone and doctests).