MSVC fix: Implement setenv() and unsetenv()
On (AppVeyor's) MSVC, setenv() and unsetenv() are not defined.
Implements simple versions of those functions, based on AGB's variant,
slightly updated and re-formatted.

Puts the functions into a separate file to not mess with the original
code too much. That way, we only have to #include that file in rtime.c
when we include updates from the original C code.

Uses the solution [1] to properly emulate snprintf() on older
MSVC platforms.

Reference:
[1]https://stackoverflow.com/questions/2915672/snprintf-and-visual-studio-2010