ci: Update appveyor builds -- template [1/2]
Increses the python version list on appveyor and removes some old build
things that are not used anymore.  Some of the tox environments that
don't exist anymore and only took up job resources.

Updates the setup using some ideas found in the tornado appveyor setup.
Removes the "bootstrap" thing, and without the bootstrap, the \Script
dir is not populated, but everything should be available normally.

Updates the allowed failures to the new environment setup.