ci: Fix osx wheel building on travis
Fixes osx image and installs and activates a virtual python3 environment
for osx on travis. Otherwise, travis defaults to python 2.7 and some of
the deploy commands fail because the packages are installed in a path
that's not in $PATH. Using a virtual environment sets all the paths
automatically and the wheel build and twine commands succeed.