setup: Update version handling and fix long desc
Uses the regex approach to extract the version number from
`src/<name>/__init__.py` instead of fixing it here.
Combines yet another source for it to reduce the chance of mishaps due
to multiple (possibly different) version numbers in the source code.

Extracts the long_description setting together with the appropriate
content type since one of the linters was complaining about that.